refactor(python): move python commands under python module
