+++ python/commands_test.go
@@ -1,4 +1,4 @@
-package main
+package python
 
 import (
 	"bytes"
