+++ scripts/test
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 go get golang.org/x/tools/cmd/cover
-go test -cover -v -coverprofile=cov.out
+go test ./... -cover -v -coverprofile=cov.out
